mux: ogg: fix overflow before widen
Fixes CID 1439601

+++ modules/mux/ogg.c
@@ -1662,7 +1662,7 @@ static int MuxBlock( sout_mux_t *p_mux, sout_input_t *p_input, block_t *p_data )
 
                 /* presentation time */
                 i_time = vlc_tick_from_samples( (p_stream->i_num_frames - 1 ) *
-                        p_stream->fmt.video.i_frame_rate_base,
+                        (int64_t) p_stream->fmt.video.i_frame_rate_base,
                         p_stream->fmt.video.i_frame_rate );
                 AddIndexEntry( p_mux, i_time, p_input );
             }
